+++ Commons/src/Commons.Application/Extensions/HttpResponseBuilder.cs
@@ -18,54 +18,9 @@ namespace MudRunner.Suspension.Application.Extensions
         public static JsonResult BuildHttpResponse<TResponseData>(this OperationResponseBase<TResponseData> response)
             where TResponseData : OperationResponseData, new()
         {
-            // BadRequest Status Code.
-            if (response.HttpStatusCode == HttpStatusCode.BadRequest)
-            {
-                response.SetBadRequestError();
-
-                return new JsonResult(response)
-                {
-                    StatusCode = (int)HttpStatusCode.BadRequest
-                };
-            }
-
-            // Unauthorized Status Code.
-            if (response.HttpStatusCode == HttpStatusCode.Unauthorized)
-            {
-                response.SetUnauthorizedError();
-
-                return new JsonResult(response)
-                {
-                    StatusCode = (int)HttpStatusCode.Unauthorized
-                };
-            }
-
-            // InternalServerError Status Code.
-            if (response.HttpStatusCode == HttpStatusCode.InternalServerError)
-            {
-                response.SetInternalServerError();
-
-                return new JsonResult(response)
-                {
-                    StatusCode = (int)HttpStatusCode.InternalServerError
-                };
-            }
-
-            // NotImplemented Status Code.
-            if (response.HttpStatusCode == HttpStatusCode.NotImplemented)
-            {
-                response.SetNotImplementedError();
-
-                return new JsonResult(response)
-                {
-                    StatusCode = (int)HttpStatusCode.NotImplemented
-                };
-            }
-
-            response.SetSuccessOk();
             return new JsonResult(response)
             {
-                StatusCode = (int)HttpStatusCode.OK
+                StatusCode = (int)response.HttpStatusCode
             };
         }
     }