+++ Suspension/src/Suspension.Core/Mapper/IMappingResolver.cs
@@ -1,6 +1,8 @@
 ﻿using MudRunner.Commons.DataContracts.Models.Profiles;
+using MudRunner.Suspension.Core.Models.NumericalMethod;
 using MudRunner.Suspension.Core.Models.SuspensionComponents;
 using MudRunner.Suspension.DataContracts.CalculateReactions;
+using MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic;
 using MudRunner.Suspension.DataContracts.RunAnalysis.Static;
 
 namespace MudRunner.Suspension.Core.Mapper
@@ -26,5 +28,12 @@ namespace MudRunner.Suspension.Core.Mapper
         /// <returns></returns>
         SuspensionSystem<TProfile> MapFrom<TProfile>(RunStaticAnalysisRequest<TProfile> runStaticAnalysisRequest, CalculateReactionsResponseData calculateReactionsResponseData)
             where TProfile : Profile;
+
+        /// <summary>
+        /// This method creates a <see cref="DynamicAnalysisResult"/> based on <see cref="NumericalMethodResult"/>.
+        /// </summary>
+        /// <param name="numericalMethodResult"></param>
+        /// <returns></returns>
+        DynamicAnalysisResult MapFrom(NumericalMethodResult numericalMethodResult);
     }
 }